Emily's initial changes to the blog
Great job on this! I love the engaging introduction and the images are so helpful for understanding how all these parts fit together. A lot of my changes were to make the content active voice and to help connect the different sections together in a clearer way. 

I think we're really close to being where we want, but I added a few notes in PR for you to consider. A few I'll call out:
* We need to add a little more context around that first image. It should probably be like the second image and explain what those different rows are for. 
* I made up a conclusion b/c the ending was abrupt. The pieces that I'm referencing are Learning Paths that we are creating to sort of guide our readers a little more down the path that we want them to take. I'll add links and descriptions to the paths/guides once I've actually created them (before next wednesday)